+++ src/main/java/com/reniec/ren/repository/CitaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.reniec.ren.modelo.Cita;
 
-public interface CitaRepository extends JpaRepository<Cita, Integer> {
+public interface CitaRepository extends JpaRepository<Cita, Long> {
     // Métodos personalizados aquí si necesitas buscar por campos de Cita
 }
